Improving code

+++ src/main/java/pt/challenge/dto/RecommendationResponse.java
@@ -2,7 +2,13 @@ package pt.challenge.dto;
 
 import java.util.List;
 
+/**
+ * Response object containing a list of recommendations and the total count.
+ *
+ * @param recommendations the list of product recommendations
+ * @param totalRecommendations the total number of recommendations in the list
+ */
 public record RecommendationResponse(
-    List<RecommendationDto> recommendations, short totalRecommendations) {
+    List<RecommendationDto> recommendations, int totalRecommendations) {
 
 }
